Initial working commit for cross entropy filter
Files changed:
- ParticleFilters.jl: Added exports for `resample_cem` and `update_cem`
- basic.jl: Added `update_cem`
- resamplers.jl: Added `resample_cem`
- sandbox.jl: Simultaneous plotting of vanilla particles and cross
  entropy particles
- Added git ignore

Caution:
This is just a minimum viable product at the moment. Resample_update is
implemented as a function belonging to the SIR filter. Should probably
be implemented as a whole new filter on its own.

Also, had to call the resample_cem throwing away predict and reweight
arguments as in the resample. Don't know how the resample works even
though its function definition does not have predict and reweight as
input arguments